Add file browser entry in webjump site
`webjump' uses completing-read to get site names, but it returns nil if
user does not input any character. Then browse-url will open the
directory of the current buffer in the system file browser, which might
be useful in some cases. I explicitly add a dummy entry to make this
special case clearer.