Fix hunspell English dictionary could not be found
(ispell-find-hunspell-dictionaries) could not parse the output of
`hunspell -D' for en_US dictionary when the environment variables LANG
or LC_ALL is not been set in Emacs. It could only find dictionary whose
name is "default". Either rename dictionary files to "default", or set
LANG or LC_ALL to en_US could resolve this issue.

If hunspell is not installed, fall back to ispell.